customer_ai/pyproject.toml
[build-system]
# setuptools>=78.1.1: CVE-2022-40897, CVE-2024-6345, CVE-2025-47273
requires = ["setuptools>=78.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ndara-customer-ai"
version = "2.0.0"
description = "AI-powered customer service system for multi-industry businesses"
readme = "README.md"
# PEP 706 (pip tar/symlink fix): 3.9.17+, 3.10.12+, 3.11.4+, or 3.12+
requires-python = ">=3.11.4"
license = {text = "Proprietary"}
authors = [
    {name = "ndara.ai"}
]
keywords = ["ai", "customer-service", "rag", "llm", "conversational-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "openai>=1.12.0",
    "langgraph>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "psycopg2-binary>=2.9.9",
    "chromadb==0.4.15",
    "numpy==1.26.4",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pydantic==2.5.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.24.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.22",
    "httpx>=0.25.2",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.1",
]

[project.urls]
Documentation = "https://github.com/ndara/ai-project"
Repository = "https://github.com/ndara/ai-project"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | chroma_db
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
